# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Soccer Connect API",
    "description": "Backend API for Soccer Connect - GTA soccer community platform",
    "version": "1.0.0",
    "contact": {
      "name": "Soccer Connect Team"
    }
  },
  "servers": [
    {
      "url": "https://soccer-connect-api.vercel.app/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:8046/api",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
  "paths": {
...
    "/messages/conversations": {
      "get": {
        "tags": ["Messages"],
        "summary": "Get user's conversations",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of conversations" }
        }
      },
      "post": {
        "tags": ["Messages"],
        "summary": "Create or get existing conversation",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "participantId": { "type": "string" },
                  "type": { "type": "string", "enum": ["direct", "group"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Conversation created or retrieved" }
        }
      }
    },
    "/messages/conversations/{id}": {
      "get": {
        "tags": ["Messages"],
        "summary": "Get single conversation with messages",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Conversation with messages" }
        }
      }
    },
    "/messages/conversations/{id}/messages": {
      "post": {
        "tags": ["Messages"],
        "summary": "Send a message",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Message sent" }
        }
      }
    },
    "/messages/conversations/{id}/poll": {
      "get": {
        "tags": ["Messages"],
        "summary": "Poll for new messages",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "New messages if any" }
        }
      }
    },
    "/messages/conversations/{id}/leave": {
      "put": {
        "tags": ["Messages"],
        "summary": "Leave a group conversation",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Left conversation" }
        }
      }
    },
    "/messages/{messageId}": {
      "delete": {
        "tags": ["Messages"],
        "summary": "Delete a message (soft delete)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "messageId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Message deleted" }
        }
      }
    },
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get user's notifications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of notifications" }
        }
      },
      "delete": {
        "tags": ["Notifications"],
        "summary": "Clear all notifications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "All notifications cleared" }
        }
      }
    },
    "/notifications/unread-count": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get count of unread notifications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Unread count" }
        }
      }
    },
    "/notifications/{id}/read": {
      "put": {
        "tags": ["Notifications"],
        "summary": "Mark notification as read",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Marked as read" }
        }
      }
    },
    "/notifications/read-all": {
      "put": {
        "tags": ["Notifications"],
        "summary": "Mark all notifications as read",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "All marked as read" }
        }
      }
    },
    "/notifications/{id}": {
      "delete": {
        "tags": ["Notifications"],
        "summary": "Delete a notification",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Notification deleted" }
        }
      }
    },
...
  }
}